+++ src/main/php/lang/ast/emit/AsymmetricVisibility.class.php
@@ -50,6 +50,7 @@ trait AsymmetricVisibility {
     } else if ($modifiers & 0x0001000) {
       $checks= [$this->private($property->name, 'modify private(set)')];
       $modifiers & MODIFIER_PRIVATE && $modifiers&= ~0x0001000;
+      $modifiers|= MODIFIER_FINAL;
     }
 
     // Emit XP meta information for the reflection API